Moved calendar sources into own folder
Added calendar ifdefs
