+++ retroshare-gui/CMakeLists.txt
@@ -62,6 +62,7 @@ option( RS_GXSCHANNELS "Enable GXS channels in GUI" ON )
 option( RS_GXSFORUMS   "Enable GXS forums in GUI"   ON )
 option( RS_GXSPOSTED   "Enable GXS posted in GUI"   ON )
 option( RS_GXSCIRCLES  "Enable GXS circles in GUI"  ON )
+option( RS_USE_CALENDAR "Build with GXS Calendar support" ON)
 option( RS_GUI_CMARK    "Enable CommonMark support in GUI"  OFF )
 set(RS_GXSIDENTITIES ON CACHE BOOL "Enable GXS identities in GUI" FORCE)
 set(RS_IDLE ON CACHE BOOL "Enable Idle support" FORCE)
