+++ retroshare-gui/src/CMakeLists.txt
@@ -115,12 +115,6 @@ list(
 	src/gui/connect/FriendRecommendDialog.cpp 
 
 	src/gui/msgs/MessagesDialog.cpp 
-	src/gui/msgs/CalendarData.cpp
-	src/gui/msgs/CalendarWidget.cpp
-	src/gui/msgs/TasksWidget.cpp
-	src/gui/msgs/CalendarPropertiesDialog.cpp
-	src/gui/msgs/EventDialog.cpp
-	src/gui/msgs/TaskDialog.cpp
 	src/gui/msgs/MessageComposer.cpp 
 	src/gui/msgs/MessageWidget.cpp 
 	src/gui/msgs/MessageWindow.cpp 
@@ -336,8 +330,6 @@ list(
 	src/gui/msgs/MessageComposer.ui 
 	src/gui/msgs/MessageWindow.ui
 	src/gui/msgs/MessageWidget.ui
-	src/gui/msgs/CalendarWidget.ui
-	src/gui/msgs/TasksWidget.ui
 
 	src/gui/settings/settingsw.ui 
 	src/gui/settings/GeneralPage.ui 
@@ -552,12 +544,6 @@ list(
 	src/gui/connect/FriendRecommendDialog.h 
 
 	src/gui/msgs/MessagesDialog.h 
-	src/gui/msgs/CalendarData.h
-	src/gui/msgs/CalendarWidget.h
-	src/gui/msgs/TasksWidget.h
-	src/gui/msgs/CalendarPropertiesDialog.h
-	src/gui/msgs/EventDialog.h
-	src/gui/msgs/TaskDialog.h
 	src/gui/msgs/MessageInterface.h 
 	src/gui/msgs/MessageComposer.h 
 	src/gui/msgs/MessageWindow.h 
@@ -724,6 +710,35 @@ if(RS_JSON_API)
 		src/gui/settings/JsonApiPage.ui
 	)
 endif(RS_JSON_API)
+
+if(RS_USE_CALENDAR)
+	add_definitions(-DRS_USE_CALENDAR)
+	list(
+		APPEND RS_GUI_SOURCES
+		src/gui/calendar/CalendarData.cpp
+		src/gui/calendar/CalendarWidget.cpp
+		src/gui/calendar/TasksWidget.cpp
+		src/gui/calendar/CalendarPropertiesDialog.cpp
+		src/gui/calendar/EventDialog.cpp
+		src/gui/calendar/TaskDialog.cpp
+	)
+
+	list(
+		APPEND RS_IMPLEMENTATION_HEADERS
+		src/gui/calendar/CalendarData.h
+		src/gui/calendar/CalendarWidget.h
+		src/gui/calendar/TasksWidget.h
+		src/gui/calendar/CalendarPropertiesDialog.h
+		src/gui/calendar/EventDialog.h
+		src/gui/calendar/TaskDialog.h
+	)
+
+	list(
+		APPEND RS_GUI_FORMS
+		src/gui/calendar/CalendarWidget.ui
+		src/gui/calendar/TasksWidget.ui
+	)
+endif(RS_USE_CALENDAR)
 
 if(RS_WEBUI)
 	list(
