+++ retroshare-gui/src/gui/calendar/CalendarWidget.ui
