+++ retroshare-gui/src/gui/calendar/EventDialog.h
@@ -22,7 +22,7 @@
 #define EVENTDIALOG_H
 
 #include <QDialog>
-#include "gui/msgs/CalendarData.h"
+#include "gui/calendar/CalendarData.h"
 
 class QComboBox;
 class QLineEdit;
