+++ retroshare-gui/src/gui/calendar/TasksWidget.ui
