+++ retroshare-gui/src/gui/msgs/MessagesDialog.h
@@ -36,8 +36,10 @@ class MessageWidget;
 class QTreeWidgetItem;
 class RsMessageModel;
 class MessageSortFilterProxyModel ;
+#ifdef RS_USE_CALENDAR
 class CalendarWidget;
 class TasksWidget;
+#endif
 
 class MessagesDialog : public MainPage
 {
@@ -112,8 +114,10 @@ private slots:
 
   void tabChanged(int tab);
   void tabCloseRequested(int tab);
+#ifdef RS_USE_CALENDAR
   void showCalendarTab();
   void showTasksTab();
+#endif
 
 private:
     void handleEvent_main_thread(std::shared_ptr<const RsEvent> event);
@@ -156,8 +160,10 @@ private:
   //RSTreeWidgetItemCompareRole *mMessageCompareRole;
 
   MessageWidget *msgWidget;
+#ifdef RS_USE_CALENDAR
   CalendarWidget *mCalendarWidget;
   TasksWidget *mTasksWidget;
+#endif
   RsMessageModel *mMessageModel;
   MessageSortFilterProxyModel *mMessageProxyModel;
